Berechtigungssteuerung

+++ src/main/kotlin/de/voegtle/wunschmanager/WishService.kt
@@ -2,6 +2,8 @@ package de.voegtle.wunschmanager
 
 import com.googlecode.objectify.Key
 import com.googlecode.objectify.ObjectifyService
+import de.voegtle.wunschmanager.util.checkOwnership
+import de.voegtle.wunschmanager.util.extractUserName
 import org.springframework.web.bind.annotation.CrossOrigin
 import org.springframework.web.bind.annotation.RequestBody
 import org.springframework.web.bind.annotation.RequestMapping
@@ -15,15 +17,24 @@ import javax.servlet.http.HttpServletRequest
 
 @RestController() class WishService {
   @CrossOrigin(origins = ["*"])
-  @RequestMapping("/wish/list") fun list(@RequestParam() list: Long): List<Wish> {
-    val tempList = ObjectifyService.ofy().load().type(Wish::class.java)
-        .ancestor(Key.create(WishList::class.java, list)).list() as MutableList<Wish>
-    return tempList
+  @RequestMapping("/wish/list") fun list(@RequestParam() list: Long, request: HttpServletRequest): List<Wish> {
+    val wishList: WishList = ObjectifyService.ofy().load().type(WishList::class.java).id(list).now()
+
+    val wishes = ObjectifyService.ofy().load().type(Wish::class.java).ancestor(wishList).list() as MutableList<Wish>
+    val userName = extractUserName(request, true)
+    if (userName == wishList.owner) {
+      wishes.forEach { it.available = it.donor == null; it.donor = null }
+    }
+    return wishes
   }
 
   @CrossOrigin(origins = ["*"])
-  @RequestMapping("/wish/create") fun create(@RequestParam() list: Long): Wish {
-    val newWish = Wish(wishList = Key.create(WishList::class.java, list))
+  @RequestMapping("/wish/create") fun create(@RequestParam() list: Long, request: HttpServletRequest): Wish {
+    val wishList: WishList = ObjectifyService.ofy().load().type(WishList::class.java).id(list).now()
+
+    checkOwnership(request, wishList, "You must be the owner of the list to create new wishes")
+
+    val newWish = Wish(wishList = Key.create(wishList))
     ObjectifyService.ofy().save().entity(newWish).now()
     return newWish
   }
@@ -31,15 +42,16 @@ import javax.servlet.http.HttpServletRequest
   @CrossOrigin(origins = ["*"])
   @RequestMapping("/wish/update", method = [RequestMethod.POST])
   fun update(@RequestBody() updateRequest: UpdateRequest, request: HttpServletRequest): Boolean {
-    val userPrincipal = request.userPrincipal
     val wish = updateRequest.wish
     wish?.let {
       val wishList: WishList = ObjectifyService.ofy().load().type(WishList::class.java).id(updateRequest.listId!!).now()
+
+      checkOwnership(request, wishList, "You must be the owner of the list to update a wishes")
+
       val existingWish = ObjectifyService.ofy().load().type(Wish::class.java).parent(wishList).id(it.id!!).now()
       existingWish.caption = it.caption
       existingWish.description = it.description
       existingWish.link = it.link
-      existingWish.available = it.available
       ObjectifyService.ofy().save().entity(existingWish).now()
       return true
     }
@@ -47,10 +59,28 @@ import javax.servlet.http.HttpServletRequest
   }
 
   @CrossOrigin(origins = ["*"])
-  @RequestMapping("/wish/delete") fun delete(@RequestParam() listId: Long, @RequestParam() wishId: Long): Boolean {
+  @RequestMapping("/wish/delete") fun delete(@RequestParam() listId: Long, @RequestParam() wishId: Long,
+                                             request: HttpServletRequest): Boolean {
     val wishList: WishList = ObjectifyService.ofy().load().type(WishList::class.java).id(listId).now()
+
+    checkOwnership(request, wishList, "You must be owner of the list to delete wishes")
+
     ObjectifyService.ofy().delete().type(Wish::class.java).parent(wishList).id(wishId).now()
     return true
   }
+
+  @CrossOrigin(origins = ["*"])
+  @RequestMapping("/wish/give") fun give(@RequestParam() listId: Long, @RequestParam() wishId: Long,
+                                         request: HttpServletRequest): Boolean {
+    val userName = extractUserName(request, true)
+
+    val wishList: WishList = ObjectifyService.ofy().load().type(WishList::class.java).id(listId).now()
+    val existingWish = ObjectifyService.ofy().load().type(Wish::class.java).parent(wishList).id(wishId).now()
+    existingWish.donor = userName
+
+    ObjectifyService.ofy().save().entity(existingWish).now()
+    return true
+  }
+
 
 }
